requirements.txt
Flask==2.3.3
requests==2.31.0
PyYAML==6.0
japanese-address-parser-py==0.1.4